backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend API gateway service for TranslatAR"
authors = [
    {name = "TranslatAR Dev Team"}
]
readme = "README.md"
requires-python = "~=3.11"
dependencies = [
    "fastapi (>=0.119.0,<0.120.0)",
    "motor (>=3.7.1,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "google-auth (>=2.42.1,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "ruff (>=0.14.1,<0.15.0)"
]

[tool.black]
line-length = 100
target-version = ['py311']
# Replace the existing [tool.ruff] and its sub-sections with this:
[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B"]

[tool.ruff.lint.isort]
# Add the 'main' modules for each specific service here
# For backend, it should be:
known-first-party = ["main", "websocket"]
# For the others, it will likely just be ["main"]
